# Implements Connection Pool Destroy Delay