--- a/include/aws/s3/private/s3_client_impl.h
+++ b/include/aws/s3/private/s3_client_impl.h
@@ -382,6 +382,8 @@ struct aws_s3_client {
         /* True if client has been flagged to finish destroying itself. Used to catch double-destroy bugs.*/
         uint32_t finish_destroy : 1;
 
+        uint32_t process_endpoint_lifecycle_changes : 1;
+
         struct aws_s3_upload_part_timeout_stats upload_part_stats;
     } synced_data;