--- a/source/s3_client.c
+++ b/source/s3_client.c
@@ -1528,6 +1528,33 @@ static void s_s3_client_process_work_default(struct aws_s3_client *client) {
             AWS_LS_S3_CLIENT, "id=%p Updating connections, assigning requests where possible.", (void *)client);
         aws_s3_client_update_connections_threaded(client);
     }
+    /*******************/
+    /* Step _: Clean up endpoints. */
+    /*******************/
+
+    /* BEGIN CRITICAL SECTION */
+    aws_s3_client_lock_synced_data(client);
+
+    for (struct aws_hash_iter iter = aws_hash_iter_begin(&client->synced_data.endpoints); !aws_hash_iter_done(&iter);
+         aws_hash_iter_next(&iter)) {
+
+        struct aws_s3_endpoint *endpoint = (struct aws_s3_endpoint *)iter.element.value;
+        AWS_LOGF_ERROR(AWS_LS_S3_ENDPOINT, "waahm7: endpoint loop:%s", aws_string_c_str(endpoint->host_name));
+
+        uint64_t now_ns = 0;
+        aws_event_loop_current_clock_time(client->process_work_event_loop, &now_ns);
+        if (endpoint->client_synced_data.run_cleanup_task && !endpoint->client_synced_data.cleanup_task_running) {
+            AWS_LOGF_ERROR(AWS_LS_S3_ENDPOINT, "waahm7: sechedling the task");
+            aws_event_loop_schedule_task_future(
+                client->process_work_event_loop,
+                endpoint->cleanup_task,
+                now_ns + aws_timestamp_convert(2, AWS_TIMESTAMP_SECS, AWS_TIMESTAMP_NANOS, NULL));
+            endpoint->client_synced_data.cleanup_task_running = true;
+        }
+    }
+
+    aws_s3_client_unlock_synced_data(client);
+    /* END CRITICAL SECTION */
 
     /*******************/
     /* Step 4: Log client stats. */