--- a/source/s3_client.c
+++ b/source/s3_client.c
@@ -1554,7 +1554,7 @@ static void s_s3_client_process_work_default(struct aws_s3_client *client) {
                         endpoint->cleanup_task,
                         now_ns +
                             aws_timestamp_convert(
-                                10 /* Value pulled from thin air */, AWS_TIMESTAMP_SECS, AWS_TIMESTAMP_NANOS, NULL));
+                                3 /* Value pulled from thin air */, AWS_TIMESTAMP_SECS, AWS_TIMESTAMP_NANOS, NULL));
                     endpoint->client_synced_data.state = AWS_S3_ENDPOINT_STATE_CLEANUP_TASK_SCHEDULED;
                 }
             } else {